deliverables_dashboard/requirements.txt
supabase-py>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
markdown>=3.5.0
